Move devtools nag into Replay Tour location (#8753)
* Staging the first part of Replay Tour

* Removing nag from replayinfo

* Linting and final wiring

* Things are mostly working

* Polishing

* Lint

* Ryan change
