+++ src/ui/utils/tour.ts
@@ -1,0 +1,5 @@
+import { Nag } from "ui/hooks/users";
+
+export function shouldShowNag(nags: Nag[], key: Nag) {
+  return nags && !nags.includes(key);
+}
